Split out MapArray into separate module (#1483)
